requirements.txt
BeautifulSoup
fogbugz==0.9.6
gitdb==0.6.4
GitPython==2.0.7
Jinja2==2.8
MarkupSafe==0.23
nosetests-json-extended==0.1.0
PyGithub==1.28
requests==2.10.0
smmap==0.9.0
wsgiref==0.1.2
sphinx
git+ssh://git@github.ibm.com/cloudant/utilitarian.git
travispy
pyvirtualdisplay
selenium
unicodecsv
